tree-wide: Run `cargo clippy --fix`
Nothing important here, just the usual spam of unnecessary `&`
basically.

+++ lib/src/cli.rs
@@ -659,7 +659,7 @@ async fn container_history(repo: &ostree::Repo, imgref: &ImageReference) -> Resu
         {
             let mut remaining = width;
             for (name, width) in columns.iter() {
-                print_column(name, *width as usize, &mut remaining);
+                print_column(name, *width, &mut remaining);
             }
             println!();
         }
